+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [unreleased]
+
+* Dropped Python 3.3 support
+* Moved from nose to pytest and updated the tests accordingly
+* Run linter on tests as well
+* Fixed several unicode related tests
+* Fixed several linter problems
+
 ## [2.7.3] - 2018-06-17
 
 * Added Makefile